# internal/app/docs/swagger.yaml
basePath: /
definitions:
  github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo:
    properties:
      limit:
        description: Límite de registros por página
        example: 10
        type: integer
      skip:
        description: Registros saltados
        example: 0
        type: integer
      total:
        description: Total de registros
        example: 100
        type: integer
      total_pages:
        description: Total de páginas
        example: 10
        type: integer
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.FieldError:
    properties:
      field:
        description: Nombre del campo con error
        example: email
        type: string
      message:
        description: Mensaje de error
        example: must be a valid email
        type: string
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.ValidationError:
    properties:
      errors:
        description: Lista de errores por campo
        items:
          $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.FieldError'
        type: array
    type: object
  internal_modules_auth.LoginDTO:
    properties:
      email:
        description: Email del usuario
        example: john@example.com
        type: string
      password:
        description: Contraseña
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  internal_modules_auth.RefreshDTO:
    properties:
      refresh_token:
        description: Refresh token
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    required:
    - refresh_token
    type: object
  internal_modules_auth.RegisterDTO:
    properties:
      email:
        description: Email del usuario
        example: john@example.com
        type: string
      name:
        description: Nombre del usuario
        example: John Doe
        minLength: 2
        type: string
      password:
        description: Contraseña (mínimo 6 caracteres)
        example: secret123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_modules_auth.TokenResponse:
    properties:
      access_token:
        description: Access token JWT
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_in:
        description: Tiempo de expiración en segundos
        example: 900
        type: integer
      refresh_token:
        description: Refresh token JWT
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  internal_modules_auth.UserInfo:
    properties:
      email:
        description: Email del usuario
        example: john@example.com
        type: string
      id:
        description: ID del usuario
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: Nombre del usuario
        example: John Doe
        type: string
    type: object
...
  internal_modules_owners.OwnerResponse:
    properties:
      address:
        type: string
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      push_tokens:
        items:
          $ref: '#/definitions/internal_modules_owners.PushTokenResponse'
        type: array
      tenant_ids:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  internal_modules_owners.PaginatedOwnersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_owners.OwnerResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo'
    type: object
  internal_modules_owners.PushTokenResponse:
    properties:
      active:
        type: boolean
      platform:
        type: string
      token:
        type: string
      updated_at:
        type: string
    type: object
  internal_modules_owners.RegisterPushTokenDTO:
    properties:
      platform:
        enum:
        - ios
        - android
        example: android
        type: string
      token:
        example: fcm-token-abc123
        type: string
    required:
    - platform
    - token
    type: object
  internal_modules_owners.UpdateOwnerDTO:
    properties:
      address:
        example: 'Calle 10 # 20-30, Medellín'
        type: string
      avatar_url:
        example: https://cdn.example.com/avatar.jpg
        type: string
      name:
        example: Juan Pérez
        type: string
      phone:
        example: +57 300 123 4567
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@vetsify.com
    name: Vetsify Support
  description: |-
    API para plataforma veterinaria multi-tenant. Soporta panel de administración (staff) y app móvil (owners).

    ## Autenticación
    Todas las rutas protegidas requieren el header `Authorization: Bearer <token>`.
    - **Staff**: Obtén tokens via `POST /api/auth/login`
    - **Owners (mobile)**: Obtén tokens via `POST /mobile/auth/login`

    ## Multi-tenancy
    Las rutas que operan sobre datos de un tenant específico (patients, species, etc.) requieren el header `X-Tenant-ID` con el ObjectID del tenant.
    Este header debe enviarse en **todas** las peticiones a rutas marcadas con el parámetro `X-Tenant-ID`.
    El frontend puede configurarlo una vez en el interceptor HTTP (axios/fetch).

    ## Roles (RBAC)
    Las rutas del panel admin (`/api/*`) están protegidas por RBAC.
    Roles disponibles: `admin`, `veterinarian`, `receptionist`, `assistant`, `accountant`.

    ## Paginación
    Las rutas que retornan listas soportan `skip` y `limit` como query params.
    Respuesta incluye `pagination: { skip, limit, total, total_pages }`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Vetsify API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario y retorna tokens
      parameters:
      - description: Credenciales
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: Credenciales inválidas
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Iniciar sesión
      tags:
      - auth
  /api/auth/me:
    get:
      consumes:
      - application/json
      description: Obtiene información del usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_auth.UserInfo'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Usuario actual
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Genera un nuevo access token usando el refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.RefreshDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Renovar token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Crea una nueva cuenta de usuario
      parameters:
      - description: Datos de registro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.RegisterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "409":
          description: Email ya existe
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar usuario
      tags:
      - auth
...
  /api/owners:
    get:
      parameters:
      - description: Skip
        in: query
        name: skip
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_owners.PaginatedOwnersResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List owners
      tags:
      - owners
  /api/owners/{id}:
    delete:
      parameters:
      - description: Owner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete owner
      tags:
      - owners
    get:
      parameters:
      - description: Owner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_owners.OwnerResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get owner by ID
      tags:
      - owners
...
  /mobile/owners/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_owners.OwnerResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get my profile
      tags:
      - mobile/owners
    patch:
      consumes:
      - application/json
      parameters:
      - description: Profile data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_owners.UpdateOwnerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_owners.OwnerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update my profile
      tags:
      - mobile/owners
  /mobile/owners/me/push-tokens:
    post:
      consumes:
      - application/json
      parameters:
      - description: Push token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_owners.RegisterPushTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Register push token
      tags:
      - mobile/owners
  /mobile/owners/me/push-tokens/{token}:
    delete:
      parameters:
      - description: Push token value
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Remove push token
      tags:
      - mobile/owners
...
securityDefinitions:
  Bearer:
    description: 'Ingresa el token con el prefijo Bearer: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"